added typesHandler to core, using it in cli and wasm-js
