+++ packages/js/core/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./combinePaths";
 export * from "./getEnvFromUriHistory";
 export * from "./is-buffer";
+export * from "./typesHandler";
